Add labels as annotations
Add all labels as annotations in image manifest.

It is important to note that the annotations will get lost if the image
is exported without OCI mediatype, or in docker format. In addition, for
the time being Dokcer engine does not support OCI annotations and
therefore pulling images with annotations in the Docker engine will
result to losing the annotations.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>